Seperation of contact page components
